# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Walker Creek Farms & Cabins API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rentals:
    get:
      operationId: listRentals
      tags: [rentals]
      summary: List all rentals
      responses:
        "200":
          description: List of rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
    post:
      operationId: createRental
      tags: [rentals]
      summary: Create a rental (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
  /rentals/{id}:
    get:
      operationId: getRental
      tags: [rentals]
      summary: Get a rental by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
        "404":
          description: Not found
    patch:
      operationId: updateRental
      tags: [rentals]
      summary: Update a rental (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalUpdate"
      responses:
        "200":
          description: Updated rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
    delete:
      operationId: deleteRental
      tags: [rentals]
      summary: Delete a rental (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rentals/{id}/photos:
    get:
      operationId: listRentalPhotos
      tags: [rentals]
      summary: List photos for a rental
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalPhoto"
    post:
      operationId: addRentalPhoto
      tags: [rentals]
      summary: Add a photo to a rental (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RentalPhotoInput"
      responses:
        "201":
          description: Photo added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalPhoto"
  /rental-photos/{id}:
    delete:
      operationId: deleteRentalPhoto
      tags: [rentals]
      summary: Delete a rental photo (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Availability ──────────────────────────────────────────────────────────────
  /rentals/{id}/availability:
    get:
      operationId: getRentalAvailability
      tags: [availability]
      summary: Get unavailable dates for a rental
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unavailable dates list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"

  # ── Bookings ─────────────────────────────────────────────────────────────────
...
  /food-items:
    get:
      operationId: listFoodItems
      tags: [food]
      summary: List all food items
      responses:
        "200":
          description: List of food items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodItem"
    post:
      operationId: createFoodItem
      tags: [food]
      summary: Create a food item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
  /food-items/{id}:
    patch:
      operationId: updateFoodItem
      tags: [food]
      summary: Update a food item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodItemUpdate"
      responses:
        "200":
          description: Updated food item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodItem"
    delete:
      operationId: deleteFoodItem
      tags: [food]
      summary: Delete a food item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── FAQ ───────────────────────────────────────────────────────────────────────
  /faq-items:
    get:
      operationId: listFaqItems
      tags: [faq]
      summary: List all FAQ items
      responses:
        "200":
          description: List of FAQ items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FaqItem"
    post:
      operationId: createFaqItem
      tags: [faq]
      summary: Create a FAQ item (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
  /faq-items/{id}:
    patch:
      operationId: updateFaqItem
      tags: [faq]
      summary: Update a FAQ item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqItemUpdate"
      responses:
        "200":
          description: Updated FAQ item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaqItem"
    delete:
      operationId: deleteFaqItem
      tags: [faq]
      summary: Delete a FAQ item (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Gift Certificates ─────────────────────────────────────────────────────────
  /gift-certificates:
    get:
      operationId: listGiftCertificates
      tags: [gift-certificates]
      summary: List all gift certificate options
      responses:
        "200":
          description: List of gift certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GiftCertificate"
    post:
      operationId: createGiftCertificate
      tags: [gift-certificates]
      summary: Create a gift certificate option (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCertificateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCertificate"
  /gift-certificates/{id}:
    patch:
      operationId: updateGiftCertificate
      tags: [gift-certificates]
      summary: Update a gift certificate (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftCertificateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GiftCertificate"
    delete:
      operationId: deleteGiftCertificate
      tags: [gift-certificates]
      summary: Delete a gift certificate option (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Blog ──────────────────────────────────────────────────────────────────────
...
  /contact-messages:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List all contact messages (admin)
      responses:
        "200":
          description: List of contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
    post:
      operationId: createContactMessage
      tags: [contact]
      summary: Submit a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /contact-messages/{id}:
    delete:
      operationId: deleteContactMessage
      tags: [contact]
      summary: Delete a contact message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Admin Dashboard ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Rental:
      type: object
      required: [id, name, type, description, guestCount, bedrooms, bathrooms, weekdayPrice, weekendPrice, cleaningFee, taxRate, isActive, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        description: { type: string }
        guestCount: { type: integer }
        bedrooms: { type: integer }
        bathrooms: { type: number }
        weekdayPrice: { type: number }
        weekendPrice: { type: number }
        cleaningFee: { type: number }
        taxRate: { type: number }
        amenities: { type: ["string", "null"] }
        coverPhoto: { type: ["string", "null"] }
        isActive: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string }
    RentalInput:
      type: object
      required: [name, type, description, guestCount, bedrooms, bathrooms, weekdayPrice, weekendPrice]
      properties:
        name: { type: string }
        type: { type: string }
        description: { type: string }
        guestCount: { type: integer }
        bedrooms: { type: integer }
        bathrooms: { type: number }
        weekdayPrice: { type: number }
        weekendPrice: { type: number }
        cleaningFee: { type: number }
        taxRate: { type: number }
        amenities: { type: string }
        coverPhoto: { type: string }
        sortOrder: { type: integer }
    RentalUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        description: { type: string }
        guestCount: { type: integer }
        bedrooms: { type: integer }
        bathrooms: { type: number }
        weekdayPrice: { type: number }
        weekendPrice: { type: number }
        cleaningFee: { type: number }
        taxRate: { type: number }
        amenities: { type: string }
        coverPhoto: { type: string }
        isActive: { type: boolean }
        sortOrder: { type: integer }
    RentalPhoto:
      type: object
      required: [id, rentalId, url, caption, sortOrder]
      properties:
        id: { type: integer }
        rentalId: { type: integer }
        url: { type: string }
        caption: { type: ["string", "null"] }
        sortOrder: { type: integer }
    RentalPhotoInput:
      type: object
      required: [url]
      properties:
        url: { type: string }
        caption: { type: string }
        sortOrder: { type: integer }
    AvailabilityResponse:
      type: object
      required: [rentalId, unavailableDates]
      properties:
        rentalId: { type: integer }
        unavailableDates: { type: array, items: { type: string } }

    # ── Booking ──────────────────────────────────────────────────────────────────
...
    FoodItem:
      type: object
      required: [id, category, name, description, price, servingSize, isAvailable, sortOrder]
      properties:
        id: { type: integer }
        category: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        servingSize: { type: ["string", "null"] }
        isAvailable: { type: boolean }
        sortOrder: { type: integer }
    FoodItemInput:
      type: object
      required: [category, name, description, price]
      properties:
        category: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        servingSize: { type: string }
        sortOrder: { type: integer }
    FoodItemUpdate:
      type: object
      properties:
        category: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        servingSize: { type: string }
        isAvailable: { type: boolean }
        sortOrder: { type: integer }

    # ── FAQ ───────────────────────────────────────────────────────────────────────
    FaqItem:
      type: object
      required: [id, question, answer, sortOrder]
      properties:
        id: { type: integer }
        question: { type: string }
        answer: { type: string }
        sortOrder: { type: integer }
    FaqItemInput:
      type: object
      required: [question, answer]
      properties:
        question: { type: string }
        answer: { type: string }
        sortOrder: { type: integer }
    FaqItemUpdate:
      type: object
      properties:
        question: { type: string }
        answer: { type: string }
        sortOrder: { type: integer }

    # ── Gift Certificates ─────────────────────────────────────────────────────────
    GiftCertificate:
      type: object
      required: [id, name, description, amount, isActive]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        amount: { type: number }
        stripeProductId: { type: ["string", "null"] }
        isActive: { type: boolean }
    GiftCertificateInput:
      type: object
      required: [name, description, amount]
      properties:
        name: { type: string }
        description: { type: string }
        amount: { type: number }
    GiftCertificateUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        amount: { type: number }
        stripeProductId: { type: string }
        isActive: { type: boolean }

    # ── Blog ──────────────────────────────────────────────────────────────────────
...
    ContactMessage:
      type: object
      required: [id, name, email, subject, message, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        subject: { type: ["string", "null"] }
        message: { type: string }
        phone: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }
    ContactMessageInput:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string }
        email: { type: string }
        subject: { type: string }
        message: { type: string }
        phone: { type: string }

    # ── Admin ──────────────────────────────────────────────────────────────────────
...