# lib/api-zod/src/generated/types/contactMessageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Walker Creek Farms & Cabins API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactMessageInput {
  name: string;
  email: string;
  subject?: string;
  message: string;
  phone?: string;
}