# lib/api-zod/src/generated/types/giftCertificate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Walker Creek Farms & Cabins API
 * OpenAPI spec version: 0.1.0
 */

export interface GiftCertificate {
  id: number;
  name: string;
  description: string;
  amount: number;
  /** @nullable */
  stripeProductId?: string | null;
  isActive: boolean;
}