# lib/api-zod/src/generated/types/rentalInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Walker Creek Farms & Cabins API
 * OpenAPI spec version: 0.1.0
 */

export interface RentalInput {
  name: string;
  type: string;
  description: string;
  guestCount: number;
  bedrooms: number;
  bathrooms: number;
  weekdayPrice: number;
  weekendPrice: number;
  cleaningFee?: number;
  taxRate?: number;
  amenities?: string;
  coverPhoto?: string;
  sortOrder?: number;
}