# lib/api-zod/src/generated/types/rentalPhotoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Walker Creek Farms & Cabins API
 * OpenAPI spec version: 0.1.0
 */

export interface RentalPhotoInput {
  url: string;
  caption?: string;
  sortOrder?: number;
}